Refactoring.

+++ common/internal/shared/hardware.go
@@ -1,7 +1,7 @@
 package shared
 
 type HardwareSpecification struct {
-	CPUCount  int   `json:"cpu"`
+	CPUCount  int64 `json:"cpu"`
 	RAMBytes  int64 `json:"ramBytes"`
 	DiskBytes int64 `json:"diskBytes"`
 }
